# src/components/docs/docSwagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is WEBSTORE API REQUIREMENT OF TRONTRON",
    "version": "1.0.5",
    "title": "Swagger WebStore",
    "contact": {
      "email": "lenguyenyenhong1199@gmail.com"
    }
  },
  "host": "http://localhost:5000",
  "basePath":"/",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/user/all": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get All User",
        "description": "This action for Staff and Owner - Get All Information of Customers",
        "operationId": "getAllUsers",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "responses": {
         
              "201": {
                "description": "successful operation",
                "schema": {
                  "type":"object",
                  "properties":{
                    "code": {
                      "type":"integer",
                      "example":"201"
                    },
                    "success":{
                      "type":"boolean",
                      "example":"true"
                    },
                    "data":{
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/User"
                      }
                    }
                  }
                }
              },
              "400":{
                "description":"Error Validate",
                "schema": {
                  "type":"object",
                  "properties":{
                    "code": {
                      "type":"integer",
                      "example":"400"
                    },
                    "success":{
                      "type":"boolean",
                      "example":"false"
                    },
                    "data":{
                      "type": "string",
                      "example":"Validate Error"
                      }
                    }
                  }
                },
             "500":{
              "description":"Server error",
              "schema": {
                "type":"object",
                "properties":{
                  "code": {
                    "type":"integer",
                    "example":"500"
                  },
                  "success":{
                    "type":"boolean",
                    "example":"false"
                  },
                  "data":{
                    "type": "string",
                    "example":"Server error"
                    }
                  }
                }
             }
             
            
          
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    },
    "/user/{userId}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Find User by ID",
        "description": "Returns a single user",
        "operationId": "getUserById",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"200"
                },
                "success":{
                  "type":"boolean",
                  "example":"true"
                },
                "data":{
                  
                    "$ref": "#/definitions/User"
               
                  
                }
              }
            }
          },
          "400":{
            "description":"Error Validate",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"400"
                },
                "success":{
                  "type":"boolean",
                  "example":"false"
                },
                "data":{
                  "type": "string",
                  "example":"Validate Error"
                  }
                }
              }
            },
         "500":{
          "description":"Server error",
          "schema": {
            "type":"object",
            "properties":{
              "code": {
                "type":"integer",
                "example":"500"
              },
              "success":{
                "type":"boolean",
                "example":"false"
              },
              "data":{
                "type": "string",
                "example":"Server error"
                }
              }
            }
         }
        },
        "security": [
          {
            "api_key": [
              
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Updates an User in the store with form data",
        "description": "",
        "operationId": "updateUserById",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user that needs to be updated",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "userName",
            "in": "formData",
            "description": "Updated name of the User",
            "required": false,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "description": "Updated email of the User",
            "required": false,
            "type": "string",
            "format":"email"
          },
          {
            "name": "address",
            "in": "formData",
            "description": "Updated address of the User",
            "required": false,
            "type": "string"
          },
          {
            "name": "phone",
            "in": "formData",
            "description": "Updated phone of the User",
            "required": false,
            "type": "string"
          },
          {
            "name": "avatar",
            "in": "formData",
            "description": "Updated avatar of the User",
            "required": false,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"200"
                },
                "success":{
                  "type":"boolean",
                  "example":"true"
                },
                "data":{
                  
                    "$ref": "#/definitions/User"             
                  
                }
              }
            }
          },
          "400":{
            "description":"Error Validate",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"400"
                },
                "success":{
                  "type":"boolean",
                  "example":"false"
                },
                "data":{
                  "type": "string",
                  "example":"Validate Error"
                  }
                }
              }
            },
         "500":{
          "description":"Server error",
          "schema": {
            "type":"object",
            "properties":{
              "code": {
                "type":"integer",
                "example":"500"
              },
              "success":{
                "type":"boolean",
                "example":"false"
              },
              "data":{
                "type": "string",
                "example":"Server error"
                }
              }
            }
         }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Deletes an User",
        "description": "Delete User By Id User",
        "operationId": "deleteUserById",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "userId",
            "in": "path",
            "description": "userId to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"200"
                },
                "success":{
                  "type":"boolean",
                  "example":"true"
                },
                "data":{
                  "type": "string",
                  "example":"Delete Successfully"                 
                }
              }
            }
          },
          "400":{
            "description":"Error Validate",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"400"
                },
                "success":{
                  "type":"boolean",
                  "example":"false"
                },
                "data":{
                  "type": "string",
                  "example":"Validate Error"
                  }
                }
              }
            },
         "500":{
          "description":"Server error",
          "schema": {
            "type":"object",
            "properties":{
              "code": {
                "type":"integer",
                "example":"500"
              },
              "success":{
                "type":"boolean",
                "example":"false"
              },
              "data":{
                "type": "string",
                "example":"Server error"
                }
              }
            }
         }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    },
    "/user/":{
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Create an User in the store with form data",
        "description": "",
        "operationId": "updateUserById",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "userName",
            "in": "formData",
            "description": " Name of the User",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "description": "email of the User",
            "required": true,
            "type": "string",
            "format":"email"
          },
          {
            "name": "address",
            "in": "formData",
            "description": "address of the User",
            "required": true,
            "type": "string"
          },
          {
            "name": "phone",
            "in": "formData",
            "description": "phone of the User",
            "required": true,
            "type": "string"
          },
          {
            "name": "avatar",
            "in": "formData",
            "description": "avatar of the User",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"200"
                },
                "success":{
                  "type":"boolean",
                  "example":"true"
                },
                "data":{
                  
                    "$ref": "#/definitions/User"             
                  
                }
              }
            }
          },
          "400":{
            "description":"Error Validate",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"400"
                },
                "success":{
                  "type":"boolean",
                  "example":"false"
                },
                "data":{
                  "type": "string",
                  "example":"Validate Error"
                  }
                }
              }
            },
         "500":{
          "description":"Server error",
          "schema": {
            "type":"object",
            "properties":{
              "code": {
                "type":"integer",
                "example":"500"
              },
              "success":{
                "type":"boolean",
                "example":"false"
              },
              "data":{
                "type": "string",
                "example":"Server error"
                }
              }
            }
         }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read:pets": "read your pets",
        "write:pets": "modify pets in your account"
      }
    }
  },
  "definitions": {
...
    "User": {
      "type": "object",
      "properties": {
        "userName": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "format":"email"
        },
        "address": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        },
        "isActive": {
          "type": "boolean",        
          "description": "User Status"
        }
      }
    },
...
  },
...
}